Separate functionality of deploying and rolling back
There are now separate modules for deploying and rolling back code (`deploy.py` and `rollback.py`).

In addition, the arg parsing and logger setup have been moved to separate files so they can be called by both the deploy and rollback scripts.

Finally, there are separate entry_points in `setup.py` for the deploy and rollback scripts.

Closes #11
