+++ setup.py
@@ -49,7 +49,8 @@ setup(
     ],
     entry_points={
         'console_scripts': [
-            "deploy = app_deployer.__main__:main"
+            'deploy = app_deployer.deploy:main',
+            'rollback = app_deployer.rollback:main',
         ]
     },
 )
